- Fixed shift logic bug (but the mono compiler still crashes)
git-svn-id: svn://localhost/mosa@532 45b04e34-05f1-4d87-a130-40a8491d5c8b

+++ Mosa/DeviceSystem/PCI/DeviceTable.cs
@@ -2001,7 +2001,7 @@ namespace Mosa.DeviceSystem.PCI
 		/// <returns></returns>
 		public static string Lookup(ushort vendorID, ushort deviceID)
 		{
-			switch ((uint)((vendorID << 16) | deviceID)) {
+			switch ((uint)(((uint)vendorID << 16) | (uint)deviceID)) {
 				case 0x00108139: return "AT-2500TX V3";
 				case 0x001C0001: return "PCAN-PCI CAN-Bus controller";
 				case 0x00700003: return "WinTV PVR-250";
@@ -9754,9 +9754,7 @@ namespace Mosa.DeviceSystem.PCI
 		/// <returns></returns>
 		public static string Lookup(ushort vendorID, ushort deviceID, ushort subSystem, ushort subVendor)
 		{
-			switch ((ulong)((vendorID << 48) | (deviceID << 32) | (subSystem << 16) | subVendor)) {
-
-// Mono has an internal compiler bug which prevents the following from compiling
+			switch ((((ulong)vendorID << 48) | ((ulong)deviceID << 32) | ((ulong)subSystem << 16) | subVendor)) {
 #if !MONO
 				case 0x0E1100460E114091: return "Smart Array 6i";
 				case 0x0E1100460E11409A: return "Smart Array 641";
